add declarations files for virgil-crypto
