+++ package.json
@@ -15,6 +15,7 @@
     "husky": "^4.0.10",
     "lerna": "^3.20.2",
     "lint-staged": "^10.0.1",
+    "rollup-plugin-copy": "^3.2.1",
     "prettier": "1.18.2"
   },
   "workspaces": [
