+++ utils/declaration.d.ts.template
@@ -1,0 +1,1 @@
+export * from './types';
